requirements.txt
fastapi
uvicorn[standard]
joblib
numpy
python-dotenv
requests
emoji
Sastrawi
nltk
scikit-learn
